apps/agent/pyproject.toml
[project]
name = "insurance-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "google-adk>=0.1.0",
  "google-genai>=0.1.0",
  "google-cloud-bigquery>=3.20.0",
  "google-cloud-secret-manager>=2.20.0",
  "google-cloud-storage>=2.16.0",
  "redis>=5.0.0",
  "pydantic>=2.7.0",
  "structlog>=24.1.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "models*", "services*"]
exclude = ["tests*", ".venv*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
